requirements.txt
Django==4.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
Pillow==8.4.0